backend/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.17
djangorestframework-simplejwt>=5.4,<5.6
gunicorn>=23,<24
psycopg[binary]>=3.2,<3.3
